refactor: rename PickProfileImgView to UploadProfileImageScreen
